requirements.txt
fastapi
uvicorn
yfinance
numpy
pandas
scikit-learn
tensorflow-cpu